Add gsheets adapter
