+++ bff/docker-compose.dev.yml
@@ -26,7 +26,7 @@ services:
     networks:
       - default
       - match-shared-network
-    command: [ "python", "-m",  "lkshmatch.website.mainWebapp" ]
+    command: [ "python", "-m",  "lkshmatch.website.main" ]
 
   # Telegram бот (опционально, можно запускать отдельно)
   match-bot:
