+++ bff/lkshmatch/adapters/gheets/gsheets.py
@@ -1,0 +1,65 @@
+from urllib.parse import urlparse, parse_qs
+import httplib2
+from googleapiclient import discovery
+from oauth2client.service_account import ServiceAccountCredentials
+
+from lkshmatch.config import settings
+
+WEBSITE_CREDENTIALS_FILE = settings.get("WEBSITE_CREDENTIALS_FILE")  # имя файла с закрытым ключом для google-таблиц
+WEBSITE_SERVICE_ACCOUNT_NAME = settings.get("WEBSITE_SERVICE_ACCOUNT_NAME")  # почта сервисного аккаунта
+credentials = ServiceAccountCredentials.from_json_keyfile_name(
+    WEBSITE_CREDENTIALS_FILE, ["https://www.googleapis.com/auth/spreadsheets", "https://www.googleapis.com/auth/drive"]
+)
+httpAuth = credentials.authorize(httplib2.Http())
+service = discovery.build("sheets", "v4", http=httpAuth)
+
+class GSheetData:
+    def __init__(self, spreadsheetId: str, sheetName: str):
+        self.spreadsheetId = spreadsheetId
+        self.sheetName = sheetName
+
+def get_sheet_data_from_url(sheet_url: str):
+    parse_result = urlparse(sheet_url)
+    sheetId = int(parse_qs(parse_result.query)["gid"][0])
+    spreadsheetId = parse_result.path.split("/")[3]
+    spreadsheet = service.spreadsheets().get(spreadsheetId=spreadsheetId).execute()
+    sheetList = spreadsheet.get("sheets")
+    sheetName = None
+    for d in sheetList:
+        if d["properties"]["sheetId"] == sheetId:
+            sheetName = d["properties"]["title"]
+            break
+    if sheetName is None:
+        raise
+    return GSheetData(spreadsheetId, sheetName)
+
+def get_data_gsheet(sheet_data: GSheetData, range: str, mod: str = "COLUMNS"):
+    # mod shoud be "ROWS" or "COLUMNS"
+    results = (
+        service.spreadsheets()
+        .values()
+        .get(spreadsheetId=sheet_data.spreadsheetId, range=sheet_data.sheetName + "!" + range, majorDimension=mod)
+        .execute()
+    )
+    return results
+
+def change_data_gsheet(sheet_data: GSheetData, range: str, data: list[list[str]], mod: str = "COLUMNS"):
+    # mod shoud be "ROWS" or "COLUMNS"
+    results = (
+        service.spreadsheets()
+        .values()
+        .batchUpdate(
+            spreadsheetId=sheet_data.spreadsheetId,
+            body={
+                "valueInputOption": "USER_ENTERED",
+                "data": [
+                    {
+                        "range": sheet_data.sheetName + "!" + range,
+                        "majorDimension": "COLUMNS",
+                        "values": data,
+                    }
+                ],
+            },
+        )
+        .execute()
+    )
