+++ bff/lkshmatch/website/main.py
@@ -1,16 +1,12 @@
 import uvicorn
-from dishka.integrations.fastapi import setup_dishka
 from fastapi import FastAPI
-from fastapi.templating import Jinja2Templates
 
 from lkshmatch.di import WEBSITE_IP, WEBSITE_PORT
-from bff.lkshmatch.website.auth.auth import auth_router
-from bff.lkshmatch.website.auth.login_middleware import LoginWallMiddleware
+from lkshmatch.website.auth.auth import auth_router
+from lkshmatch.website.auth.login_middleware import LoginWallMiddleware
 from lkshmatch.website.activities import activities_router
-from lkshmatch.website.table_adapter import table_adapter_router
+from lkshmatch.website.gsheets import table_adapter_router
 
-
-templates = Jinja2Templates("./lkshmatch/website/templates")
 app = FastAPI()
 
 app.include_router(auth_router, prefix="/auth")
